+++ eng/tox/mypy_hard_failure_packages.py
@@ -9,5 +9,6 @@ MYPY_HARD_FAILURE_OPTED = [
   "azure-core",
   "azure-eventhub",
   "azure-ai-textanalytics",
-  "azure-ai-formrecognizer"
+  "azure-ai-formrecognizer",
+  "azure-storage-blob",
 ]
