+++ sdk/storage/azure-storage-blob/azure/storage/blob/_container_client.py
@@ -7,7 +7,7 @@
 
 import functools
 from typing import (  # pylint: disable=unused-import
-    Union, Optional, Any, Iterable, AnyStr, Dict, List, Tuple, IO, Iterator,
+    Union, Optional, Any, Iterable, AnyStr, Dict, List, Tuple, IO, Iterator, cast,
     TYPE_CHECKING
 )
 
@@ -23,19 +23,19 @@ from azure.core.paging import ItemPaged
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.pipeline import Pipeline
 from azure.core.pipeline.transport import HttpRequest
+from azure.core.pipeline.policies import HTTPPolicy, SansIOHTTPPolicy # pylint: disable=ungrouped-imports
 
-from ._shared.base_client import StorageAccountHostsMixin, TransportWrapper, parse_connection_str, parse_query
+from ._shared.base_client import TransportWrapper, parse_connection_str
 from ._shared.request_handlers import add_metadata_headers, serialize_iso
 from ._shared.response_handlers import (
     process_storage_error,
     return_response_headers,
     return_headers_and_deserialized)
-from ._generated import AzureBlobStorage, VERSION
 from ._generated.models import (
     StorageErrorException,
     SignedIdentifier)
 from ._deserialize import deserialize_container_properties
-from ._serialize import get_modify_conditions, get_container_cpk_scope_info, get_api_version
+from ._serialize import get_modify_conditions, get_container_cpk_scope_info
 from ._models import ( # pylint: disable=unused-import
     ContainerProperties,
     BlobProperties,
@@ -44,17 +44,19 @@ from ._models import ( # pylint: disable=unused-import
     BlobPrefix)
 from ._lease import BlobLeaseClient, get_access_conditions
 from ._blob_client import BlobClient
+from ._container_client_base import ContainerClientBase
 
 if TYPE_CHECKING:
     from azure.core.pipeline.transport import HttpTransport, HttpResponse  # pylint: disable=ungrouped-imports
-    from azure.core.pipeline.policies import HTTPPolicy # pylint: disable=ungrouped-imports
     from datetime import datetime
+    from ._download import StorageStreamDownloader
     from ._models import (  # pylint: disable=unused-import
         PublicAccess,
         AccessPolicy,
         ContentSettings,
         StandardBlobTier,
         PremiumPageBlobTier)
+PoliciesType = List[Union[HTTPPolicy, SansIOHTTPPolicy]]
 
 
 def _get_blob_name(blob):
@@ -69,7 +71,7 @@ def _get_blob_name(blob):
         return blob
 
 
-class ContainerClient(StorageAccountHostsMixin):
+class ContainerClient(ContainerClientBase):
     """A client to interact with a specific container, although that container
     may not yet exist.
 
@@ -125,31 +127,6 @@ class ContainerClient(StorageAccountHostsMixin):
             :dedent: 8
             :caption: Creating the container client directly.
     """
-    def __init__(
-            self, account_url,  # type: str
-            container_name,  # type: str
-            credential=None,  # type: Optional[Any]
-            **kwargs  # type: Any
-        ):
-        # type: (...) -> None
-        try:
-            if not account_url.lower().startswith('http'):
-                account_url = "https://" + account_url
-        except AttributeError:
-            raise ValueError("Container URL must be a string.")
-        parsed_url = urlparse(account_url.rstrip('/'))
-        if not container_name:
-            raise ValueError("Please specify a container name.")
-        if not parsed_url.netloc:
-            raise ValueError("Invalid URL: {}".format(account_url))
-
-        _, sas_token = parse_query(parsed_url.query)
-        self.container_name = container_name
-        self._query_str, credential = self._format_query_string(sas_token, credential)
-        super(ContainerClient, self).__init__(parsed_url, service='blob', credential=credential, **kwargs)
-        self._client = AzureBlobStorage(self.url, pipeline=self._pipeline)
-        self._client._config.version = get_api_version(kwargs, VERSION)  # pylint: disable=protected-access
-
     def _format_url(self, hostname):
         container_name = self.container_name
         if isinstance(container_name, six.text_type):
@@ -241,7 +218,7 @@ class ContainerClient(StorageAccountHostsMixin):
 
     @distributed_trace
     def create_container(self, metadata=None, public_access=None, **kwargs):
-        # type: (Optional[Dict[str, str]], Optional[Union[PublicAccess, str]], **Any) -> None
+        # type: (Optional[Dict[str, str]], Optional[Union[PublicAccess, str]], **Any) -> bool
         """
         Creates a new container under the specified account. If the container
         with the same name already exists, the operation fails.
@@ -274,10 +251,10 @@ class ContainerClient(StorageAccountHostsMixin):
         """
         headers = kwargs.pop('headers', {})
         timeout = kwargs.pop('timeout', None)
-        headers.update(add_metadata_headers(metadata)) # type: ignore
+        headers.update(add_metadata_headers(metadata))
         container_cpk_scope_info = get_container_cpk_scope_info(kwargs)
         try:
-            return self._client.container.create( # type: ignore
+            ret_val = self._client.container.create(
                 timeout=timeout,
                 access=public_access,
                 container_cpk_scope_info=container_cpk_scope_info,
@@ -285,7 +262,9 @@ class ContainerClient(StorageAccountHostsMixin):
                 headers=headers,
                 **kwargs)
         except StorageErrorException as error:
+            ret_val = None
             process_storage_error(error)
+        return cast(bool, ret_val)
 
     @distributed_trace
     def delete_container(
@@ -393,7 +372,7 @@ class ContainerClient(StorageAccountHostsMixin):
                 :dedent: 8
                 :caption: Acquiring a lease on the container.
         """
-        lease = BlobLeaseClient(self, lease_id=lease_id) # type: ignore
+        lease = BlobLeaseClient(self, lease_id=lease_id)
         kwargs.setdefault('merge_span', True)
         timeout = kwargs.pop('timeout', None)
         lease.acquire(lease_duration=lease_duration, timeout=timeout, **kwargs)
@@ -411,9 +390,10 @@ class ContainerClient(StorageAccountHostsMixin):
         :rtype: dict(str, str)
         """
         try:
-            return self._client.container.get_account_info(cls=return_response_headers, **kwargs) # type: ignore
+            ret_val = self._client.container.get_account_info(cls=return_response_headers, **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
+        return cast(Dict, ret_val)
 
     @distributed_trace
     def get_container_properties(self, **kwargs):
@@ -449,12 +429,13 @@ class ContainerClient(StorageAccountHostsMixin):
                 cls=deserialize_container_properties,
                 **kwargs)
         except StorageErrorException as error:
+            response = None
             process_storage_error(error)
         response.name = self.container_name
-        return response # type: ignore
+        return cast(ContainerProperties, response)
 
     @distributed_trace
-    def set_container_metadata( # type: ignore
+    def set_container_metadata(
             self, metadata=None,  # type: Optional[Dict[str, str]]
             **kwargs
         ):
@@ -508,7 +489,7 @@ class ContainerClient(StorageAccountHostsMixin):
         mod_conditions = get_modify_conditions(kwargs)
         timeout = kwargs.pop('timeout', None)
         try:
-            return self._client.container.set_metadata( # type: ignore
+            ret_val = self._client.container.set_metadata(
                 timeout=timeout,
                 lease_access_conditions=access_conditions,
                 modified_access_conditions=mod_conditions,
@@ -516,7 +497,9 @@ class ContainerClient(StorageAccountHostsMixin):
                 headers=headers,
                 **kwargs)
         except StorageErrorException as error:
+            ret_val = None
             process_storage_error(error)
+        return cast(Dict, ret_val)
 
     @distributed_trace
     def get_container_access_policy(self, **kwargs):
@@ -614,15 +597,14 @@ class ContainerClient(StorageAccountHostsMixin):
             if value:
                 value.start = serialize_iso(value.start)
                 value.expiry = serialize_iso(value.expiry)
-            identifiers.append(SignedIdentifier(id=key, access_policy=value)) # type: ignore
-        signed_identifiers = identifiers # type: ignore
+            identifiers.append(SignedIdentifier(id=key, access_policy=value))
         lease = kwargs.pop('lease', None)
         mod_conditions = get_modify_conditions(kwargs)
         access_conditions = get_access_conditions(lease)
         timeout = kwargs.pop('timeout', None)
         try:
-            return self._client.container.set_access_policy(
-                container_acl=signed_identifiers or None,
+            ret_val = self._client.container.set_access_policy(
+                container_acl=identifiers or None,
                 timeout=timeout,
                 access=public_access,
                 lease_access_conditions=access_conditions,
@@ -631,6 +613,7 @@ class ContainerClient(StorageAccountHostsMixin):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
+        return cast(Dict, ret_val)
 
     @distributed_trace
     def list_blobs(self, name_starts_with=None, include=None, **kwargs):
@@ -896,10 +879,10 @@ class ContainerClient(StorageAccountHostsMixin):
             The timeout parameter is expressed in seconds.
         :rtype: None
         """
-        blob_client = self.get_blob_client(blob) # type: ignore
+        blob_client = self.get_blob_client(blob)
         kwargs.setdefault('merge_span', True)
         timeout = kwargs.pop('timeout', None)
-        blob_client.delete_blob( # type: ignore
+        blob_client.delete_blob(
             delete_snapshots=delete_snapshots,
             timeout=timeout,
             **kwargs)
@@ -967,71 +950,9 @@ class ContainerClient(StorageAccountHostsMixin):
         :returns: A streaming object (StorageStreamDownloader)
         :rtype: ~azure.storage.blob.StorageStreamDownloader
         """
-        blob_client = self.get_blob_client(blob) # type: ignore
+        blob_client = self.get_blob_client(blob)
         kwargs.setdefault('merge_span', True)
         return blob_client.download_blob(offset=offset, length=length, **kwargs)
-
-    def _generate_delete_blobs_options(
-        self, snapshot=None,
-        delete_snapshots=None,
-        request_id=None,
-        lease_access_conditions=None,
-        modified_access_conditions=None,
-        **kwargs
-    ):
-        """This code is a copy from _generated.
-
-        Once Autorest is able to provide request preparation this code should be removed.
-        """
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
-        if_modified_since = None
-        if modified_access_conditions is not None:
-            if_modified_since = modified_access_conditions.if_modified_since
-        if_unmodified_since = None
-        if modified_access_conditions is not None:
-            if_unmodified_since = modified_access_conditions.if_unmodified_since
-        if_match = None
-        if modified_access_conditions is not None:
-            if_match = modified_access_conditions.if_match
-        if_none_match = None
-        if modified_access_conditions is not None:
-            if_none_match = modified_access_conditions.if_none_match
-
-        # Construct parameters
-        timeout = kwargs.pop('timeout', None)
-        query_parameters = {}
-        if snapshot is not None:
-            query_parameters['snapshot'] = self._client._serialize.query("snapshot", snapshot, 'str')  # pylint: disable=protected-access
-        if timeout is not None:
-            query_parameters['timeout'] = self._client._serialize.query("timeout", timeout, 'int', minimum=0)  # pylint: disable=protected-access
-
-        # Construct headers
-        header_parameters = {}
-        if delete_snapshots is not None:
-            header_parameters['x-ms-delete-snapshots'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "delete_snapshots", delete_snapshots, 'DeleteSnapshotsOptionType')
-        if request_id is not None:
-            header_parameters['x-ms-client-request-id'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "request_id", request_id, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "lease_id", lease_id, 'str')
-        if if_modified_since is not None:
-            header_parameters['If-Modified-Since'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "if_modified_since", if_modified_since, 'rfc-1123')
-        if if_unmodified_since is not None:
-            header_parameters['If-Unmodified-Since'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "if_unmodified_since", if_unmodified_since, 'rfc-1123')
-        if if_match is not None:
-            header_parameters['If-Match'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "if_match", if_match, 'str')
-        if if_none_match is not None:
-            header_parameters['If-None-Match'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "if_none_match", if_none_match, 'str')
-
-        return query_parameters, header_parameters
 
     @distributed_trace
     def delete_blobs(self, *blobs, **kwargs):
@@ -1116,39 +1037,6 @@ class ContainerClient(StorageAccountHostsMixin):
             reqs.append(req)
 
         return self._batch_send(*reqs, **options)
-
-    def _generate_set_tier_options(
-        self, tier, rehydrate_priority=None, request_id=None, lease_access_conditions=None, **kwargs
-    ):
-        """This code is a copy from _generated.
-
-        Once Autorest is able to provide request preparation this code should be removed.
-        """
-        lease_id = None
-        if lease_access_conditions is not None:
-            lease_id = lease_access_conditions.lease_id
-
-        comp = "tier"
-        timeout = kwargs.pop('timeout', None)
-        # Construct parameters
-        query_parameters = {}
-        if timeout is not None:
-            query_parameters['timeout'] = self._client._serialize.query("timeout", timeout, 'int', minimum=0)  # pylint: disable=protected-access
-        query_parameters['comp'] = self._client._serialize.query("comp", comp, 'str')  # pylint: disable=protected-access, specify-parameter-names-in-call
-
-        # Construct headers
-        header_parameters = {}
-        header_parameters['x-ms-access-tier'] = self._client._serialize.header("tier", tier, 'str')  # pylint: disable=protected-access, specify-parameter-names-in-call
-        if rehydrate_priority is not None:
-            header_parameters['x-ms-rehydrate-priority'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "rehydrate_priority", rehydrate_priority, 'str')
-        if request_id is not None:
-            header_parameters['x-ms-client-request-id'] = self._client._serialize.header(  # pylint: disable=protected-access
-                "request_id", request_id, 'str')
-        if lease_id is not None:
-            header_parameters['x-ms-lease-id'] = self._client._serialize.header("lease_id", lease_id, 'str')  # pylint: disable=protected-access
-
-        return query_parameters, header_parameters
 
     @distributed_trace
     def set_standard_blob_tier_blobs(
@@ -1302,8 +1190,8 @@ class ContainerClient(StorageAccountHostsMixin):
         blob_name = _get_blob_name(blob)
         _pipeline = Pipeline(
             transport=TransportWrapper(self._pipeline._transport), # pylint: disable = protected-access
-            policies=self._pipeline._impl_policies # pylint: disable = protected-access
-        )
+            policies=cast(PoliciesType, self._pipeline._impl_policies) # pylint: disable = protected-access
+        ) # type: Pipeline
         return BlobClient(
             self.url, container_name=self.container_name, blob_name=blob_name, snapshot=snapshot,
             credential=self.credential, api_version=self.api_version, _configuration=self._config,
