+++ sdk/storage/azure-storage-blob/azure/storage/blob/_deserialize.py
@@ -49,13 +49,13 @@ def deserialize_container_properties(response, obj, headers):
 
 def get_page_ranges_result(ranges):
     # type: (PageList) -> Tuple[List[Dict[str, int]], List[Dict[str, int]]]
-    page_range = [] # type: ignore
+    page_range = []
     clear_range = [] # type: List
     if ranges.page_range:
-        page_range = [{'start': b.start, 'end': b.end} for b in ranges.page_range] # type: ignore
+        page_range = [{'start': b.start, 'end': b.end} for b in ranges.page_range]
     if ranges.clear_range:
         clear_range = [{'start': b.start, 'end': b.end} for b in ranges.clear_range]
-    return page_range, clear_range  # type: ignore
+    return page_range, clear_range
 
 
 def service_stats_deserialize(generated):
