+++ sdk/storage/azure-storage-blob/azure/storage/blob/_shared/base_client.py
@@ -13,6 +13,8 @@ from typing import (  # pylint: disable=unused-import
     List,
     Type,
     Tuple,
+    Iterator,
+    cast,
     TYPE_CHECKING,
 )
 import logging
@@ -28,7 +30,10 @@ import six
 from azure.core.configuration import Configuration
 from azure.core.exceptions import HttpResponseError
 from azure.core.pipeline import Pipeline
-from azure.core.pipeline.transport import RequestsTransport, HttpTransport
+from azure.core.pipeline.transport import (
+    RequestsTransport,
+    HttpTransport,
+)
 from azure.core.pipeline.policies import (
     RedirectPolicy,
     ContentDecodePolicy,
@@ -57,6 +62,11 @@ from .._version import VERSION
 from .._generated.models import StorageErrorException
 from .response_handlers import process_storage_error, PartialBatchErrorException
 
+if TYPE_CHECKING:
+    from azure.core.pipeline.transport import (
+        HttpResponse,
+        HttpRequest
+    )
 
 _LOGGER = logging.getLogger(__name__)
 _SERVICE_PARAMS = {
@@ -65,7 +75,7 @@ _SERVICE_PARAMS = {
     "file": {"primary": "FileEndpoint", "secondary": "FileSecondaryEndpoint"},
     "dfs": {"primary": "BlobEndpoint", "secondary": "BlobEndpoint"},
 }
-
+T = Optional[Union[BearerTokenCredentialPolicy, SharedKeyCredentialPolicy]]
 
 class StorageAccountHostsMixin(object):  # pylint: disable=too-many-instance-attributes
     def __init__(
@@ -213,7 +223,7 @@ class StorageAccountHostsMixin(object):  # pylint: disable=too-many-instance-att
 
     def _create_pipeline(self, credential, **kwargs):
         # type: (Any, **Any) -> Tuple[Configuration, Pipeline]
-        self._credential_policy = None
+        self._credential_policy = None # type: T
         if hasattr(credential, "get_token"):
             self._credential_policy = BearerTokenCredentialPolicy(credential, STORAGE_OAUTH_SCOPE)
         elif isinstance(credential, SharedKeyCredentialPolicy):
@@ -224,11 +234,11 @@ class StorageAccountHostsMixin(object):  # pylint: disable=too-many-instance-att
         config = kwargs.get("_configuration") or create_configuration(**kwargs)
         if kwargs.get("_pipeline"):
             return config, kwargs["_pipeline"]
-        config.transport = kwargs.get("transport")  # type: ignore
+        _transport = kwargs.get("transport")
         kwargs.setdefault("connection_timeout", CONNECTION_TIMEOUT)
         kwargs.setdefault("read_timeout", READ_TIMEOUT)
-        if not config.transport:
-            config.transport = RequestsTransport(**kwargs)
+        if not _transport:
+            _transport = RequestsTransport(**kwargs)
         policies = [
             QueueMessagePolicy(),
             config.headers_policy,
@@ -246,17 +256,18 @@ class StorageAccountHostsMixin(object):  # pylint: disable=too-many-instance-att
             DistributedTracingPolicy(**kwargs),
             HttpLoggingPolicy(**kwargs)
         ]
-        return config, Pipeline(config.transport, policies=policies)
+        return config, Pipeline(_transport, policies=policies)
 
     def _batch_send(
-        self, *reqs,  # type: HttpRequest
-        **kwargs
+        self, *reqs, # type: HttpRequest
+        **kwargs # type: Any
     ):
+        # type: (...) -> Iterator[HttpResponse]
         """Given a series of request, do a Storage batch call.
         """
         # Pop it here, so requests doesn't feel bad about additional kwarg
         raise_on_any_failure = kwargs.pop("raise_on_any_failure", True)
-        request = self._client._client.post(  # pylint: disable=protected-access
+        request = self._client._client.post(  # type: ignore # pylint: disable=protected-access
             url='https://{}/?comp=batch'.format(self.primary_hostname),
             headers={
                 'x-ms-version': self.api_version
@@ -290,9 +301,10 @@ class StorageAccountHostsMixin(object):  # pylint: disable=too-many-instance-att
                     )
                     raise error
                 return iter(parts)
-            return parts
+            return cast(Iterator, parts)
         except StorageErrorException as error:
             process_storage_error(error)
+        return cast(Iterator, None)
 
 class TransportWrapper(HttpTransport):
     """Wrapper class that ensures that an inner client created
@@ -386,24 +398,29 @@ def create_configuration(**kwargs):
     config.logging_policy = StorageLoggingPolicy(**kwargs)
     config.proxy_policy = ProxyPolicy(**kwargs)
 
+    # Would be good for Configuration to support __setattr__ as per this cheat sheet
+    # https://mypy.readthedocs.io/en/latest/cheat_sheet_py3.html#when-you-re-puzzled-or-when-things-are-complicated
     # Storage settings
-    config.max_single_put_size = kwargs.get("max_single_put_size", 64 * 1024 * 1024)
-    config.copy_polling_interval = 15
+    setattr(config, 'max_single_put_size', kwargs.get("max_single_put_size", 64 * 1024 * 1024))
+    setattr(config, 'copy_polling_interval', 15)
 
     # Block blob uploads
-    config.max_block_size = kwargs.get("max_block_size", 4 * 1024 * 1024)
-    config.min_large_block_upload_threshold = kwargs.get("min_large_block_upload_threshold", 4 * 1024 * 1024 + 1)
-    config.use_byte_buffer = kwargs.get("use_byte_buffer", False)
+    setattr(config, 'max_block_size', kwargs.get("max_block_size", 4 * 1024 * 1024))
+    setattr(config,
+        'min_large_block_upload_threshold',
+        kwargs.get("min_large_block_upload_threshold", 4 * 1024 * 1024 + 1)
+    )
+    setattr(config, 'use_byte_buffer', kwargs.get("use_byte_buffer", False))
 
     # Page blob uploads
-    config.max_page_size = kwargs.get("max_page_size", 4 * 1024 * 1024)
+    setattr(config, 'max_page_size', kwargs.get("max_page_size", 4 * 1024 * 1024))
 
     # Blob downloads
-    config.max_single_get_size = kwargs.get("max_single_get_size", 32 * 1024 * 1024)
-    config.max_chunk_get_size = kwargs.get("max_chunk_get_size", 4 * 1024 * 1024)
+    setattr(config, 'max_single_get_size', kwargs.get("max_single_get_size", 32 * 1024 * 1024))
+    setattr(config, 'max_chunk_get_size', kwargs.get("max_chunk_get_size", 4 * 1024 * 1024))
 
     # File uploads
-    config.max_range_size = kwargs.get("max_range_size", 4 * 1024 * 1024)
+    setattr(config, 'max_range_size', kwargs.get("max_range_size", 4 * 1024 * 1024))
     return config
