+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/_lease_async.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 
 from typing import (  # pylint: disable=unused-import
-    Union, Optional, Any, IO, Iterable, AnyStr, Dict, List, Tuple,
+    Union, Optional, Any, IO, Iterable, AnyStr, Dict, List, Tuple, cast,
     TypeVar, TYPE_CHECKING
 )
 
@@ -16,7 +16,7 @@ from .._generated.models import (
     StorageErrorException,
     LeaseAccessConditions)
 from .._serialize import get_modify_conditions
-from .._lease import BlobLeaseClient as LeaseClientBase
+from .._lease_client_base import BlobLeaseClientBase
 
 if TYPE_CHECKING:
     from datetime import datetime
@@ -25,7 +25,7 @@ if TYPE_CHECKING:
     ContainerClient = TypeVar("ContainerClient")
 
 
-class BlobLeaseClient(LeaseClientBase):
+class BlobLeaseClient(BlobLeaseClientBase):
     """Creates a new BlobLeaseClient.
 
     This client provides lease operations on a BlobClient or ContainerClient.
@@ -106,9 +106,9 @@ class BlobLeaseClient(LeaseClientBase):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
-        self.id = response.get('lease_id')  # type: str
-        self.last_modified = response.get('last_modified')   # type: datetime
-        self.etag = kwargs.get('etag')  # type: str
+        self.id = response.get('lease_id')
+        self.last_modified = response.get('last_modified')
+        self.etag = kwargs.get('etag')
 
     @distributed_trace_async
     async def renew(self, **kwargs):
@@ -152,9 +152,9 @@ class BlobLeaseClient(LeaseClientBase):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
-        self.etag = response.get('etag')  # type: str
-        self.id = response.get('lease_id')  # type: str
-        self.last_modified = response.get('last_modified')   # type: datetime
+        self.etag = response.get('etag')
+        self.id = response.get('lease_id')
+        self.last_modified = response.get('last_modified')
 
     @distributed_trace_async
     async def release(self, **kwargs):
@@ -196,9 +196,9 @@ class BlobLeaseClient(LeaseClientBase):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
-        self.etag = response.get('etag')  # type: str
-        self.id = response.get('lease_id')  # type: str
-        self.last_modified = response.get('last_modified')   # type: datetime
+        self.etag = response.get('etag')
+        self.id = response.get('lease_id')
+        self.last_modified = response.get('last_modified')
 
     @distributed_trace_async
     async def change(self, proposed_lease_id, **kwargs):
@@ -240,9 +240,9 @@ class BlobLeaseClient(LeaseClientBase):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
-        self.etag = response.get('etag')  # type: str
-        self.id = response.get('lease_id')  # type: str
-        self.last_modified = response.get('last_modified')   # type: datetime
+        self.etag = response.get('etag')
+        self.id = response.get('lease_id')
+        self.last_modified = response.get('last_modified')
 
     @distributed_trace_async
     async def break_lease(self, lease_break_period=None, **kwargs):
@@ -293,4 +293,4 @@ class BlobLeaseClient(LeaseClientBase):
                 **kwargs)
         except StorageErrorException as error:
             process_storage_error(error)
-        return response.get('lease_time') # type: ignore
+        return cast(int, response.get('lease_time'))
